Continue moving to common job service
