+++ totaleconomy-bukkit/src/test/java/com/ericgrandt/totaleconomy/listeners/PlayerListenerTest.java
@@ -12,13 +12,13 @@ import com.ericgrandt.totaleconomy.common.data.AccountData;
 import com.ericgrandt.totaleconomy.common.data.BalanceData;
 import com.ericgrandt.totaleconomy.common.data.CurrencyData;
 import com.ericgrandt.totaleconomy.common.data.Database;
-import com.ericgrandt.totaleconomy.common.data.JobDataOld;
+import com.ericgrandt.totaleconomy.common.data.JobData;
 import com.ericgrandt.totaleconomy.common.data.dto.AccountDto;
 import com.ericgrandt.totaleconomy.common.data.dto.BalanceDto;
 import com.ericgrandt.totaleconomy.common.data.dto.JobExperienceDto;
 import com.ericgrandt.totaleconomy.common.econ.CommonEconomy;
+import com.ericgrandt.totaleconomy.common.services.JobService;
 import com.ericgrandt.totaleconomy.commonimpl.BukkitLogger;
-import com.ericgrandt.totaleconomy.services.JobService;
 import com.zaxxer.hikari.HikariDataSource;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -78,7 +78,7 @@ public class PlayerListenerTest {
         when(databaseMock.getDataSource()).thenReturn(mock(HikariDataSource.class));
         when(databaseMock.getDataSource().getConnection()).then(x -> TestUtils.getConnection());
 
-        JobDataOld jobData = new JobDataOld(databaseMock);
+        JobData jobData = new JobData(new BukkitLogger(loggerMock), databaseMock);
         AccountData accountData = new AccountData(databaseMock);
         BalanceData balanceData = new BalanceData(databaseMock);
         CurrencyData currencyData = new CurrencyData(databaseMock);
@@ -89,7 +89,7 @@ public class PlayerListenerTest {
             balanceData,
             currencyData
         );
-        JobService jobServiceMock = new JobService(loggerMock, jobData);
+        JobService jobServiceMock = new JobService(jobData);
         PlayerListener sut = new PlayerListener(economy, jobServiceMock, null);
 
         // Act
